Use bit specific, not ABI types, for floating point assembly.
Use int32_t/int64_t rather than long int, long long int, etc.
This makes the code somewhat more portable if ever we support,
for example, ilp32 on RV64.

+++ sysdeps/riscv/rv32/rvd/s_lrint.c
@@ -18,11 +18,12 @@
 
 #include <math.h>
 #include <libm-alias-double.h>
+#include <stdint.h>
 
 long int
 __lrint (double x)
 {
-  long int res;
+  int32_t res;
   asm ("fcvt.w.d %0, %1" : "=r" (res) : "f" (x));
   return res;
 }
